Refactor template library modules
